Init ui part using vue.js
This commits adds all the structure of the UI part, it is a vue.js
app, so it has all the backbone of an standard vue app, including
the package.json with the dependencies, the configuration files in
order to build the software, and the tree folder structure where
the components and files are going to be.

This start point contains the main component (RaistlinUI.vue), a
component for the tasks view (Tasks.vue) and the base html
(index.html).

Signed-off-by: David Moreno <dmorenolumb@gmail.com>