Add tasks/list petition in django in order to get the tasks list
This commit adds a new url resource (/tasks/) in order to group
all the petitions related to tasks actions in one sub-django-app
called tasks.

This tasks app has been created with the /tasks/lists GET
petition, this petition returns a JSON with the tasks of arthur
or an error (in JSON format as well) notifying if something went
wrong.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>