Show tasks list in the Tasks vue component
This commit modifies the component Tasks of the UI vue part in
order to show the tasks list of the arthur server, formatting the
response into a table.

If the server returns an empty list of tasks, a warning message
will appear. And if there is an error in the petition, an error
message with the status and the content message will appear as
well.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>