[ui] Add config of a hot development server between vue and Django
This commit adds the configuration part in order to have a hot
development server for deploy the frontend part queriying the
server (Django) part.

Specifically, it configures in the devServer a proxy server that
will redirect all the petitions to the Django url server (in our
case localhost:8000) instead to the webpack dev server.

This feature just work for the development server (yarn server),
not in production.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>