[ui] Add Task component to see the task details
This commits adds a new component for the UI part, the component
is Task that shows by default the list of jobs that it has. The
component calls the server in order to retrieve the data of the
task, doing a petition to the url tasks/id/<task_id>. The Django
server will return a JSON (or an error) with the data.

It shows the jobs in a table with the status and the ID of the job.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>