[ui] Show Task details and jobs as list in the task page
This commit improves the task main page showing the task summary
as a card and a color that shows in a visual way, the status
of the task. Moreover, at the bottom there are the jobs, showing
them as a list with the color that shows the job status as well.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>