[ui] Add Job view in the Task page
This commit adds a new route to see the jobs separated from the
task, it will show in the task page, a div at the bottom showing
the log, and a new item in the tabs menu (next to the "Jobs" tab,
separated by an angle icon) that is the job id and is selected by
default.

Finally, the "Jobs" tab works in order to return to the jobs list.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>