[ui] Show tasks as a list with the CSS improved
This commit changes the table that shows the tasks list for a
CSS friendly list with the same information that the table
had.

In order to reuse some logic between components, a "mixins"
folder has been created with a file that has the logic related to
the task list item (just the methods that define the CSS).

Signed-off-by: David Moreno <dmorenolumb@gmail.com>