[ui] Add form in order to add a new task
This commit adds a new component that is the form that will be used
to add task to the KingArthur server, this form has all the data
and it will be sent to the backend part in order to add it to the
KingArthur server.

This commits adds also a new route in the ui part, and a button in
the Tasks list page in order to access to the form.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>