[tasks] Add view in order to add tasks to the KingArthur server
This commit adds the url and the view class with all the logic
needed that will parse the data sent by the ui part and then
will send the petition to add the task to the KingArthur server.

Moreover, if the response of the KingArthur server does not return
200, a message with an error will be returned to the UI part.

As we don't have users for now, it is needed to add a csrf exempt
in order to make POST petitions, this will be eliminated when we
add the users management.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>