[ui] Show last 10 jobs with colored squares in the Tasks view
In order to have an overview of the last executed jobs of a Task,
this commit adds 10 colored squares on the Tasks view (on the
right of each Task card) that show the status of the last 10 jobs
of the Task.

Moreover, there is a tooltip on each square that shows the ID of
the job and the status written.

Signed-off-by: David Moreno <dmorenolumb@gmail.com>