[ui] Fix Task/Job rendering when refreshing or going back/forward
This commit fixes the bug when the Task page is refreshed or the
user clicks in the going back/forward arrows of the browser.
It uses a watch in the URL (as it does with the task_id) for
the job_id param. If changes without clicking the button, it
needs to render the job.

Signed-off-by: David Moreno <dmoreno@bitergia.com>