[ui] Show job result when its status is finished
In order to complete the information of the job in the Job view,
this commit adds a card that shows the basic info of the job
(number an id) and if the job has result (when it finished), the
result will be showed in the card as well.

The log div is under the card now.

Signed-off-by: David Moreno <dmoreno@bitergia.com>